requirements.txt
python-dotenv
requests
flask
aiohttp
dateparser
pycryptodome
ujson
websockets
asn1crypto==1.4.0
black==21.9b0
cbor2==5.5.0
cffi==1.15.0
click==8.0.3
cryptography==41.0.7
mccabe==0.6.1
mypy==1.4.1
mypy-extensions==1.0.0
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
pyOpenSSL==23.3.0
regex==2021.10.8
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing_extensions==4.9.0